+++ src/main/java/com/ezzenix/chatanimation/impl/ModMenuImpl.java
@@ -1,7 +1,7 @@
 //? if fabric {
 package com.ezzenix.chatanimation.impl;
 
-import com.ezzenix.chatanimation.config.ModConfigScreen;
+import com.ezzenix.chatanimation.lib.config.ConfigScreen;
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
 
@@ -9,7 +9,7 @@ public class ModMenuImpl implements ModMenuApi {
 
     @Override
     public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return ModConfigScreen::create;
+		return ConfigScreen::new;
     }
 
 }