+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.3.0
+
+- New config screen.
+- Opacity uses time instead of ticks. No longer limited to 20 fps.
+
 ## 1.2.0
 
 - Opacity fading for all versions